Split login logic
